requirements.txt
torch #==2.1.2
torchvision #==0.16.2
transformers #==4.40.2
sentencepiece #==0.1.99
decord==0.6.0
Pillow==10.1.0
timm

Pillow==10.1.0
torch #==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
transformers==4.44.2
librosa==0.9.0
soundfile==0.12.1
vector-quantize-pytorch==1.18.5
vocos==0.1.0
decord
moviepy

#flash_attn

#git+https://github.com/gbvolkov/AIAssistantsLibPackage.git
